+++ src/main/java/com/gol/ants_quests/hibernate/repositories/QuestsRepository.java
@@ -2,9 +2,10 @@ package com.gol.ants_quests.hibernate.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+
 import com.gol.ants_quests.hibernate.entities.Quest;
 
 @Repository
-public interface QuestsRepository extends JpaRepository<Quest, Integer> {
+public interface QuestsRepository extends JpaRepository<Quest, Long> {
 
 }
